Add tests and upgrade index
