+++ Dockerfile
@@ -2,7 +2,8 @@ FROM python:3.10-slim
 
 WORKDIR /app
 
-COPY . /app
+COPY app/main.py /app/
+COPY storage/ /app/
 
 RUN pip install -r requirements.txt
